@@ -6,6 +6,14 @@ from haystack_integrations.components.embedders.voyage_embedders.voyage_contextu
     VoyageContextualizedDocumentEmbedder,
 )
 from haystack_integrations.components.embedders.voyage_embedders.voyage_document_embedder import VoyageDocumentEmbedder
+from haystack_integrations.components.embedders.voyage_embedders.voyage_multimodal_embedder import (
+    VoyageMultimodalEmbedder,
+)
 from haystack_integrations.components.embedders.voyage_embedders.voyage_text_embedder import VoyageTextEmbedder
 
-__all__ = ["VoyageContextualizedDocumentEmbedder", "VoyageDocumentEmbedder", "VoyageTextEmbedder"]
+__all__ = [
+    "VoyageContextualizedDocumentEmbedder",
+    "VoyageDocumentEmbedder",
+    "VoyageMultimodalEmbedder",
+    "VoyageTextEmbedder",
+]
